requirements.txt
openai>=1.1.0
python-dotenv
requests>=2.32.0